once more with clamped feeling
