+++ hypothesis-python/src/hypothesis/strategies/_internal/utils.py
@@ -17,8 +17,7 @@ import attr
 
 from hypothesis.internal.cache import LRUReusedCache
 from hypothesis.internal.compat import dataclass_asdict
-from hypothesis.internal.conjecture.junkdrawer import clamp
-from hypothesis.internal.floats import float_to_int
+from hypothesis.internal.floats import clamp, float_to_int
 from hypothesis.internal.reflection import proxies
 from hypothesis.vendor.pretty import pretty
