+++ hypothesis-python/tests/nocover/test_strategy_state.py
@@ -15,7 +15,7 @@ from random import Random
 from hypothesis import Verbosity, assume, settings
 from hypothesis.database import InMemoryExampleDatabase
 from hypothesis.internal.compat import PYPY
-from hypothesis.internal.floats import float_to_int, int_to_float, is_negative
+from hypothesis.internal.floats import clamp, float_to_int, int_to_float, is_negative
 from hypothesis.stateful import Bundle, RuleBasedStateMachine, rule
 from hypothesis.strategies import (
     binary,
@@ -35,18 +35,6 @@ from hypothesis.strategies import (
 )
 
 AVERAGE_LIST_LENGTH = 2
-
-
-def clamp(lower, value, upper):
-    """Given a value and optional lower/upper bounds, 'clamp' the value so that
-    it satisfies lower <= value <= upper."""
-    if (lower is not None) and (upper is not None) and (lower > upper):
-        raise ValueError(f"Cannot clamp with lower > upper: {lower!r} > {upper!r}")
-    if lower is not None:
-        value = max(lower, value)
-    if upper is not None:
-        value = min(value, upper)
-    return value
 
 
 class HypothesisSpec(RuleBasedStateMachine):
